+++ tests/bearer_auth.rs
@@ -1,11 +1,15 @@
 use cometd_client::types::{access_token::Bearer, AccessToken};
+use reqwest::header::{HeaderMap, HeaderName, HeaderValue};
 
 #[test]
 fn test_basic_auth() {
-    let bearer = Bearer::new("Vasya");
+    let bearer = Bearer::new("Vasya").unwrap();
 
     assert_eq!(
         bearer.get_authorization_header(),
-        [("authorization", "Bearer Vasya".into())]
+        HeaderMap::from_iter([(
+            HeaderName::from_static("authorization"),
+            HeaderValue::from_static("Bearer Vasya")
+        )])
     );
 }